+++ app/src/androidTest/java/com/kodingan/instagrammockup/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.kodingan.uicollection
+package com.kodingan.instagrammockup
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.kodingan.uicollection", appContext.packageName)
+        assertEquals("com.kodingan.instagrammockup", appContext.packageName)
     }
 }
