+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">UI Collection</string>
+    <string name="app_name">Instagram Mockup</string>
     <string name="username">Username</string>
     <string name="_1">1</string>
     <string name="hosico_cat">hosico_cat</string>
