+++ app/src/main/res/values/themes.xml
@@ -1,11 +1,11 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.UICollection" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.InstagramMockup" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.UICollection" parent="Base.Theme.UICollection" />
+    <style name="Theme.InstagramMockup" parent="Base.Theme.InstagramMockup" />
 
     <style name="Circular">
         <item name="cornerSize">50%</item>
