+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "UI Collection"
+rootProject.name = "Instagram Mockup"
 include ':app'
